HZN-1520: Rework geocoder configuration persistence
